+++ CHANGELOG.md
@@ -7,7 +7,9 @@ and this project adheres to [Semantic Versioning](http://semver.org/spec/v2.0.0.
 
 ## [Unreleased]
 
-<!-- ### Added-->
+### Added
+
+- Added search feature for overview page [#347](https://github.com/openkfw/TruBudget/issues/347)
 
 ### Changed